chore: replace a function
